Controller/View Relationship
Able to setup the content, Title of tab/page, and data sent through a
controller. The content is pulled from a view file called
v_users_profile. The data define in the controller c_users.php and sent
to the view file v_users_profile.php to be included in the content. The
title of the page is defined in the 'profile' method of this controller.
All of this is done with the following lines of code within the
'profile' method of the controller c_users.php:
#Set up the View
$this->template->content =  View::instance('v_users_profile');
$this->template->title = "Profile";

#Pass the data to the View
$this->template->content->user_name = $user_name;